rename gossip4 module to gossip
