+++ cmd/node.go
@@ -6,30 +6,30 @@ import (
 
 	"github.com/AsynkronIT/protoactor-go/actor"
 	g3types "github.com/quorumcontrol/tupelo-go-sdk/gossip3/types"
-	"github.com/quorumcontrol/tupelo-go-sdk/gossip4/types"
+	"github.com/quorumcontrol/tupelo-go-sdk/gossip/types"
 	"github.com/quorumcontrol/tupelo-go-sdk/p2p"
 	"github.com/quorumcontrol/tupelo-go-sdk/tracing"
 	"github.com/spf13/cobra"
 
 	"github.com/quorumcontrol/tupelo/gossip3to4"
-	"github.com/quorumcontrol/tupelo/gossip4"
+	"github.com/quorumcontrol/tupelo/gossip"
 	"github.com/quorumcontrol/tupelo/nodebuilder"
 )
 
-func runGossip4Node(ctx context.Context, config *nodebuilder.Config, group *types.NotaryGroup) (*actor.PID, error) {
+func runGossipNode(ctx context.Context, config *nodebuilder.Config, group *types.NotaryGroup) (*actor.PID, error) {
 	p2pNode, peer, err := p2p.NewHostAndBitSwapPeer(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("error creating p2p node: %v", err)
 	}
 
-	nodeCfg := &gossip4.NewNodeOptions{
+	nodeCfg := &gossip.NewNodeOptions{
 		P2PNode:     p2pNode,
 		SignKey:     config.PrivateKeySet.SignKey,
 		NotaryGroup: group,
 		DagStore:    peer,
 	}
 
-	node, err := gossip4.NewNode(ctx, nodeCfg)
+	node, err := gossip.NewNode(ctx, nodeCfg)
 	if err != nil {
 		return nil, fmt.Errorf("error creating new node: %v", err)
 	}
@@ -47,7 +47,7 @@ func runGossip4Node(ctx context.Context, config *nodebuilder.Config, group *type
 	return node.PID(), nil
 }
 
-func runGossip3To4Node(ctx context.Context, group *g3types.NotaryGroup, gossip4PID *actor.PID) error {
+func runGossip3To4Node(ctx context.Context, group *g3types.NotaryGroup, gossipPID *actor.PID) error {
 	p2pNode, _, err := p2p.NewHostAndBitSwapPeer(ctx)
 	if err != nil {
 		return fmt.Errorf("error creating p2p node: %v", err)
@@ -56,7 +56,7 @@ func runGossip3To4Node(ctx context.Context, group *g3types.NotaryGroup, gossip4P
 	node := gossip3to4.NewNode(ctx, &gossip3to4.NodeConfig{
 		P2PNode:     p2pNode,
 		NotaryGroup: group,
-		Gossip4Node: gossip4PID,
+		GossipNode: gossipPID,
 	})
 
 	node.Start(ctx)
@@ -66,7 +66,7 @@ func runGossip3To4Node(ctx context.Context, group *g3types.NotaryGroup, gossip4P
 
 var nodeCmd = &cobra.Command{
 	Use:   "node",
-	Short: "Run a tupelo node (gossip4)",
+	Short: "Run a tupelo node (gossip)",
 	Long:  ``,
 	Run: func(cmd *cobra.Command, args []string) {
 		ctx, cancel := context.WithCancel(context.Background())
@@ -103,13 +103,13 @@ var nodeCmd = &cobra.Command{
 		// get the gossip4 notary group
 		localKeys := config.PrivateKeySet
 		localSigner := types.NewLocalSigner(&localKeys.DestKey.PublicKey, localKeys.SignKey)
-		gossip4NotaryGroup, err := config.NotaryGroupConfig.NotaryGroup(localSigner)
+		gossipNotaryGroup, err := config.NotaryGroupConfig.NotaryGroup(localSigner)
 		if err != nil {
 			panic(fmt.Errorf("error generating notary group: %v", err))
 		}
 
 		// spin up a gossip4 node
-		pid, err := runGossip4Node(ctx, config, gossip4NotaryGroup)
+		pid, err := runGossipNode(ctx, config, gossipNotaryGroup)
 		if err != nil {
 			panic(err)
 		}
