+++ go.mod
@@ -32,6 +32,8 @@ require (
 	github.com/shibukawa/configdir v0.0.0-20170330084843-e180dbdc8da0
 	github.com/spf13/cobra v0.0.5
 	github.com/stretchr/testify v1.3.0
+	go.uber.org/atomic v1.5.1 // indirect
+	go.uber.org/multierr v1.4.0 // indirect
 	golang.org/x/crypto v0.0.0-20191111213947-16651526fdb4
 	google.golang.org/grpc v1.21.1
 )
