+++ nodebuilder/config.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/quorumcontrol/tupelo-go-sdk/bls"
 	g3types "github.com/quorumcontrol/tupelo-go-sdk/gossip3/types"
-	"github.com/quorumcontrol/tupelo-go-sdk/gossip4/types"
+	"github.com/quorumcontrol/tupelo-go-sdk/gossip/types"
 )
 
 type PrivateKeySet struct {
