+++ nodebuilder/nodebuilder.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/quorumcontrol/tupelo-go-sdk/gossip4/middleware"
-	"github.com/quorumcontrol/tupelo-go-sdk/gossip4/types"
+	"github.com/quorumcontrol/tupelo-go-sdk/gossip/middleware"
+	"github.com/quorumcontrol/tupelo-go-sdk/gossip/types"
 	"github.com/quorumcontrol/tupelo-go-sdk/tracing"
 
 	"github.com/AsynkronIT/protoactor-go/actor"
