+++ nodebuilder/nodebuilder_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/quorumcontrol/tupelo-go-sdk/p2p"
 
-	"github.com/quorumcontrol/tupelo-go-sdk/gossip4/types"
+	"github.com/quorumcontrol/tupelo-go-sdk/gossip/types"
 
 	"github.com/quorumcontrol/tupelo/testnotarygroup"
